+++ activerecord/lib/active_record/connection_adapters/sqlite3/schema_creation.rb
@@ -5,6 +5,14 @@ module ActiveRecord
     module SQLite3
       class SchemaCreation < SchemaCreation # :nodoc:
         private
+          def visit_TableDefinition(o)
+            create_sql = super
+            if ConnectionAdapters::SQLite3Adapter.strict_tables && o.strict_table
+              create_sql << " STRICT"
+            end
+            create_sql
+          end
+
           def supports_index_using?
             false
           end